+++ .gitignore
@@ -6,4 +6,5 @@ build/
 gen/
 debugger/lib
 exampleProject
+testData
 .DS_Store
