+++ src/test/kotlin/org/rustSlowTests/CargoProjectResolveTest.kt
@@ -1,24 +1,11 @@
 package org.rustSlowTests
 
-import com.intellij.openapi.application.runWriteAction
 import com.intellij.openapi.util.SystemInfo
-import com.intellij.openapi.vfs.LocalFileSystem
-import com.intellij.openapi.vfs.VfsUtil
-import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiManager
-import com.intellij.psi.PsiRecursiveElementVisitor
 import org.rust.FileTreeBuilder
 import org.rust.TestProject
 import org.rust.cargo.RustWithToolchainTestBase
-import org.rust.cargo.project.settings.toolchain
-import org.rust.cargo.project.workspace.CargoProjectWorkspaceService
-import org.rust.cargo.project.workspace.cargoWorkspace
 import org.rust.fileTree
-import org.rust.lang.core.psi.RsFile
 import org.rust.lang.core.psi.RsPath
-import org.rust.utils.fullyRefreshDirectory
-import kotlin.system.measureTimeMillis
 
 class CargoProjectResolveTest : RustWithToolchainTestBase() {
 
@@ -261,49 +248,4 @@ class CargoProjectResolveTest : RustWithToolchainTestBase() {
         fileTree { builder() }.create(project, project.baseDir).apply {
             refreshWorkspace()
         }
-
-    private fun refreshWorkspace() {
-        CargoProjectWorkspaceService.getInstance(module).syncUpdate(module.project.toolchain!!)
-        if (module.cargoWorkspace == null) {
-            error("Failed to update a test Cargo project")
-        }
-    }
-
-
-    @Suppress("unused")
-    fun measureResolveTime() {
-        val pathToCargoSource = "/home/user/examples/cargo"
-        val base = openRealProject(pathToCargoSource)
-        val toml = base.findFileByRelativePath("src/cargo/util/toml.rs")
-            ?: error("failed to find toml file")
-        val psiManager = PsiManager.getInstance(project)
-        val psi = psiManager.findFile(toml) as RsFile
-
-        val elapsed = measureTimeMillis {
-            psi.accept(object : PsiRecursiveElementVisitor() {
-                override fun visitElement(element: PsiElement) {
-                    super.visitElement(element)
-                    element.reference?.resolve()
-                }
-            })
-        }
-
-        println("elapsed = ${elapsed}ms")
-    }
-
-    @Suppress("unused")
-    private fun openRealProject(path: String): VirtualFile {
-        runWriteAction {
-            VfsUtil.copyDirectory(
-                this,
-                LocalFileSystem.getInstance().findFileByPath(path)!!,
-                project.baseDir,
-                { true }
-            )
-            fullyRefreshDirectory(project.baseDir)
-        }
-
-        refreshWorkspace()
-        return project.baseDir
-    }
 }
